💥 Harmonize entity unique ids
They were previously based on the name meaning they were unique but not
necessarily fixed in time (when changing a typo for instance).

We are now using the "id" provided by heishamon which is less likely to
change because it is just an id provided when a new message is added.

Change-Id: I6e05b677ae563a5316fbabdddbb7030ceaef05cc