🐛 Only react to markers of 3.1/3.2 when relevant
Before this patch, a firmware 3.3 (or any custom build) would alternate
between 3.2 and the other version. It would happen because
stats_firmware_contain_version would be set to true and we would receive
markers of version 3.2 as well and update HA state.

Now we ignore mzarkers of 3.1/3.2 until we are sure stats topic does not
contain version information

Fix #152