🐛 Improve integration precision
This is an attempt to improve situation for #168.

Before this patch, trapezoid method used to "hallucinate" energy usage:
after a period of 1h at 0 W, a new value at 800W would be considered as
400Wh instead of 0.

After this patch, the same scenario should count 0Wh.

We still have an issue for cases when consumption/production does not
change for a long time (and is different from 0) but it should not be a
problem because heatpump usually vary in consumption/production quite
often.