Add "External Compressor Contral Switch"
Add the actual "switch" for "External Compressor Control"
The currently existing "External Compressor Control" only enables that setting being controllable but does not have any physical effect on the heat pump. Only if both are "ON", there is an action.
Tested for L series but should be available also for earlier series with "optional PCB emulation" enabled.